Image/image_evaluation_framework/pyproject.toml
[project]
name = "image-evaluation-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform>=1.108.0",
    "google-cloud-storage>=2.19.0",
    "google-genai>=1.29.0",
    "ipycanvas>=0.14.1",
    "ipykernel>=6.30.1",
    "ipython>=9.4.0",
    "ipywidgets>=8.1.7",
    "langfuse>=3.3.2",
    "matplotlib>=3.10.5",
    "pydantic>=2.11.7",
    "streamlit>=1.49.1",
    "streamlit-extras>=0.7.7",
]

